[feat]: split datatypes by lang of valList
The initial idea was to split the values by the ident attribute of of the tei:dataSpec-element. This seems to be a bad idea. Instead now a lang-attribute is used on the vallist, which is used for filtering, when the docs are rendered. At the moment this is just used for rendering the possible content of elements.