[ci/cd]: setup pre-commit hooks
This commit adds some hooks to the project. This hooks will be executed before a commit is applied to the repo. The hook, which checks the xml-format can also be executed via the `Taskfile`. So just use `run fmt` or `run fmt-xml` to format the xml-files. The hooks need to be installed at first – just use the `run install` command to install all project dependencies (including `pre-commit`) as well as the hooks.