[fix]: autoreload while `run serve`
There was a problem with the liveserver, which produced an infinite reload loop.
This was caused by some kind of a race-condition between the file
watcher of the server and the hook, which generated the `.md`-files for
each element-specification. This fix removes the docs-directory from the
watch-list – only directories listed in `mkdocs.yml` will be watched!
Note: A change on `index.md` will no longer trigger a reload, because
only directories can be watched.