[refac]: project (folder) structure
The changes introduced in this commit are responsible for a complete
reorganization of the ssrq-schema project – with a goal of a much
clearer structure. The `src`-folder now contains just files that are the
source of the schema as well as of the docs-page. All things, like
python- or xslt-scripts, which are used to process those files are
stored inside `utils`. The TEI-stylsheets submodule is also embedded
here.