[docs/feat]: plugin/hook for creating relative links
This plugins does nearly the same as a lot of the plugins listed in https://github.com/mkdocs/catalog#-links--references. Most of these plugins are not working you could expect it or they are not function complete. The custom plugin / hook should do the work for our purpose. It was developed with TDD in mind. See the tests dir for more details.