feat: changed datatypes of witness-ids + wit-attr
The id of a `<witness>`-element is now a custom datatype, which can be
referenced by the values of the attribute `@wit`. The `@wit`-attribute
no longer uses the `IDREFS`-datatype, because it needs to support a
fragment identifier style (for parts) at the end. The
schematron-constraint for the attribute will ensure the correct usage of
everythin inside the attribute.